requirements.txt
spotipy
musicbrainzngs
python-dotenv
networkx
matplotlib